rename de  columna a tablero.
