+++ src/Elementos/tablero.tsx
